+++ src/GeoIp2/Model/AbstractModel.php
@@ -23,7 +23,15 @@ abstract class AbstractModel implements \JsonSerializable
      */
     protected function get($field)
     {
-        return isset($this->raw[$field]) ? $this->raw[$field] : null;
+        if (isset($this->raw[$field])) {
+            return $this->raw[$field];
+        } else {
+            if (preg_match('/^is_/', $field)) {
+                return false;
+            } else {
+                return null;
+            }
+        }
     }
 
     /**