+++ src/GeoIp2/Record/AbstractRecord.php
@@ -25,7 +25,11 @@ abstract class AbstractRecord implements \JsonSerializable
         if ($this->__isset($attr)) {
             return $this->record[$key];
         } elseif ($this->validAttribute($attr)) {
-            return null;
+            if (preg_match('/^is_/', $key)) {
+                return false;
+            } else {
+                return null;
+            }
         } else {
             throw new \RuntimeException("Unknown attribute: $attr");
         }