+++ CHANGELOG.md
@@ -9,6 +9,8 @@ CHANGELOG
 * Added support for the GeoIP2 Anonymous IP database. The
   `GeoIP2\Database\Reader` class now has an `anonymousIp` method which returns
   a `GeoIP2\Model\AnonymousIp` object.
+* Boolean attributes like those in the `GeoIP2\Record\Traits` class now return
+ `false` insteadof `null` when they were not true.
 
 2.0.0 (2014-09-22)
 ------------------